Chapter 7 challenges done
Added support for mixed type comparison expression ( 'string'/'double'  '>'/'>='/'<'/'<=' ) And also for Strings comparison, this evaluates using the length of the string as the operand and comparing it to the number.
Added support for concatenation between string and number. So every time you have a string as operand for the + operator, you cast both operands to strings.
Hanlding division by zero